Crate pattern_finder

Example on tic-tac-toe

We are going to find one of these patterns:

by horizontally
X X X
_ _ _
_ _ _

by vertically
X _ _
X _ _
X _ _

by diagonally from the left
X _ _
_ X _
_ _ X

by diagonally from the right
_ _ X
_ X _
X _ _

use pattern_finder::prelude::*;

enum Mark {
    X,
    O
}

use Mark::*;

let board = [
    [X, X, O],
    [O, X, O],
    [X, X, X]
];

// сonvert `board` to a one-dimensional array
let board = board.iter().flatten().collect::<Vec<_>>();

let horizontal_pattern = PatternConstructor::new(&board)
    .items_per_once(3)
    .pairs(|prev, curr| prev == curr);

// by calling the `transposed` method
// we will iterate over the columns as if they were rows,
// and over the rows as if they were columns
let vertical_pattern = PatternConstructor::new(&board)
    .items_per_once(3)
    .transposed()
    .pairs(|prev, curr| prev == curr);

// `offset` - offset at each iteration step
// `iterate_on_both_sides` - iterate simultaneously from both sides
let diagonal_pattern = PatternConstructor::new(&board)
    .items_per_once(3)
    .offset(|i| i)
    .pairs(|prev, curr| prev == curr)
    .iterate_on_both_sides();

let pattern_finder = PatternFinder::new(board)
    .add_pattern(horizontal_pattern)
    .add_pattern(vertical_pattern)
    .add_pattern(diagonal_pattern);

Structs

PatternFinder